LDPC_Decode using floats instead of doubles (for decode, some utilities still doubles). This version runs on stm32 with good results but uses too much memory. No data on times yet.

+++ codec2-dev/src/mpdecode_core_test.h
@@ -35,9 +35,9 @@ void encode(struct LDPC *ldpc, unsigned char ibits[], unsigned char pbits[]);
 
 void ldpc_alloc_mem(struct LDPC *ldpc);
 void ldpc_free_mem(struct LDPC *ldpc);
-int run_ldpc_decoder(struct LDPC *ldpc, char out_char[], double input[], int *parityCheckCount);
+int run_ldpc_decoder(struct LDPC *ldpc, char out_char[], float input[], int *parityCheckCount);
 
-void sd_to_llr(double llr[], double sd[], int n);
+void sd_to_llr(float llr[], double sd[], int n);
 
 void Demod2D(double symbol_likelihood[], COMP r[], COMP S_matrix[], float EsNo, float fading[], float mean_amp, int number_symbols);
 void Somap(double bit_likelihood[], double symbol_likelihood[], int number_symbols);
